Add some talk-name access in permissions for screenshot on Wayland.
It is actually unclear to me if "talk-name" is about the bus name or an
interface name. It actually looks like the bus name is expected, but if
so, it would just give a whole bunch of useless permissions at once.

In Freedesktop for instance, the bus name is
org.freedesktop.portal.Desktop, which also gives access to camera,
location, remote desktop API, and much more. If this is what is
expected, then it is a bit illogical with the targetted permission
concept.
So let's first try to only allow specific interface access then test if
it works.

See: https://gitlab.gnome.org/GNOME/gimp/-/issues/6106